backend/requirements.txt
google-adk==1.15.1
fastapi
uvicorn[standard]
python-dotenv
python-multipart